[FIX] builtins removed from Python 3
* Reverse wrapper courtesy of @rco-odoo's original P3 branch
* thin compat module stripped down from werkzeug

issue 8530

+++ odoo/addons/base/ir/ir_ui_menu.py
@@ -133,9 +133,9 @@ class IrUiMenu(models.Model):
             if not self._context.get('ir.ui.menu.full_list'):
                 menus = menus._filter_visible_menus()
             if offset:
-                menus = menus[long(offset):]
+                menus = menus[offset:]
             if limit:
-                menus = menus[:long(limit)]
+                menus = menus[:limit]
         return len(menus) if count else menus
 
     @api.multi
